bfelf_loader: implement relocations for aarch64
